Rename file tutorial to be about MFS
- Rename tutorial according to new focus on the Mutable File System
- Update routes and index listings accordingly
