+++ src/main.js
@@ -26,10 +26,10 @@ import LessonDataStructures02 from './tutorials/Data-Structures/02.vue'
 import LessonDataStructures03 from './tutorials/Data-Structures/03.vue'
 import LessonDataStructures04 from './tutorials/Data-Structures/04.vue'
 import LessonDataStructures05 from './tutorials/Data-Structures/05.vue'
-import LessonFileBasics01 from './tutorials/File-Basics/01.vue'
-import LessonFileBasics02 from './tutorials/File-Basics/02.vue'
-import LessonFileBasics03 from './tutorials/File-Basics/03.vue'
-import LessonFileBasics04 from './tutorials/File-Basics/04.vue'
+import MutableFileSystem01 from './tutorials/Mutable-File-System/01.vue'
+import MutableFileSystem02 from './tutorials/Mutable-File-System/02.vue'
+import MutableFileSystem03 from './tutorials/Mutable-File-System/03.vue'
+import MutableFileSystem04 from './tutorials/Mutable-File-System/04.vue'
 
 Vue.use(VueRouter)
 
@@ -55,10 +55,10 @@ const routes = [
   { path: '/blog/05', component: LessonBlog05 },
   { path: '/blog/06', component: LessonBlog06 },
   { path: '/blog/07', component: LessonBlog07 },
-  { path: '/file-basics/01', component: LessonFileBasics01 },
-  { path: '/file-basics/02', component: LessonFileBasics02 },
-  { path: '/file-basics/03', component: LessonFileBasics03 },
-  { path: '/file-basics/04', component: LessonFileBasics04 },
+  { path: '/mutable-file-system/01', component: MutableFileSystem01 },
+  { path: '/mutable-file-system/02', component: MutableFileSystem02 },
+  { path: '/mutable-file-system/03', component: MutableFileSystem03 },
+  { path: '/mutable-file-system/04', component: MutableFileSystem04 },
   { path: '*', name: '404' }
 ]
