+++ src/tutorials/Mutable-File-System/04.vue
@@ -53,6 +53,8 @@ const validate = async (result, ipfs) => {
   // but it also has a native error msg showing that's unclear
 }
 
+//blah
+
 
 
 const code = `const run = async (files) => {
